Update EfModels    &    Add IAdressAdderOrEditorFormOpener ,AdressAdderOrEditorFormOpener to FormOpeners Of AdressFormHelpers    &    Impliment IAdressAdderOrEditorFormOepner  to AdderOrEditorForm    &     Add AdressDataAdders to DataAdders Of PublicHelpers    &    Add AdressDataBuilders to DataBuilders Of PublicHelpers    &    Add AdressesData Geter to DataGeters Of PublicHelpers    &    Impliment  IAdressesDataGeter,IAdressDataBuilders,IAdressDataAdders  to AdressAdderOrEditorForm

+++ StationeryStoreDataLayer/Models/AdressTable.cs
@@ -19,7 +19,7 @@ public partial class AdressTable
 
     public string HouseNumber { get; set; } = null!;
 
-    public int PostalCode { get; set; }
+    public long PostalCode { get; set; }
 
     public virtual UserTable User { get; set; } = null!;
 }
